genomecomb.c fix: no long long in fprintf with PRId64 format

+++ generic/genomecomb.c
@@ -108,7 +108,7 @@ genomecomb_tsv_select_ObjCmd (ClientData clientData,	Tcl_Interp *interp, int arg
 		if (verbose < 0) {
 			totalread += line->size+1;
 			if (totalread >= next) {
-				fprintf(stderr,"%" PRId64 "\n",(long long)totalread);
+				fprintf(stderr,"%" PRId64 "\n",totalread);
 				next += abs(verbose);
 			}
 		} else if (verbose != 0 && line_nr >= next) {
@@ -196,7 +196,7 @@ genomecomb_tsv_select_ObjCmd (ClientData clientData,	Tcl_Interp *interp, int arg
 	if (line) {DStringDestroy(line);}
 	if (array) {DStringArrayDestroy(array);}
 	if (verbose < 0) {
-		fprintf(stderr,"%" PRId64 "\n",(long long)totalread);
+		fprintf(stderr,"%" PRId64 "\n",totalread);
 	} else if (verbose > 0) {
 		fprintf(stderr,"%d\n",line_nr);
 	}
